refactor(public_app): Split large files into modular structure
- api_views.py: Extract docx and utils modules
- tasks.py: Split into tasks/ directory
- tool_repo_concat_api.py: Extract clone, concat, utils
- pages.py: Extract data, donate, shortcuts
- views/status/api.py: Split into api/ directory
- views/tools.py: Extract data modules (image, pdf, text, other)
- templates/about.html: Update structure
