+++ apps/public_app/api_views.py
@@ -1,147 +1,40 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-# Timestamp: "2025-12-11 (ywatanabe)"
-# File: /home/ywatanabe/proj/scitex-cloud/apps/public_app/api_views.py
-# ----------------------------------------
 """
 API views for public_app tools.
 
 Provides backend functionality for browser-based tools that require
 server-side processing.
+
+Re-exports from specialized submodules:
+- api_utils: Helper functions for file handling
+- api_docx: DOCX to LaTeX conversion
 """
 
-import base64
 import json
 import logging
+import os
+import re
+import shutil
 import tempfile
 from pathlib import Path
-from typing import Any, Dict, Optional
 
 from django.http import JsonResponse
 from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_http_methods
 
+from .api_docx import docx2tex_convert
+from .api_utils import (
+    detect_bundle_type,
+    get_bundle_dimensions_from_png,
+    get_svg_dimensions,
+    read_bundle_metadata,
+)
+
 logger = logging.getLogger("scitex")
 
-# ----------------------------------------
-
-
-def _detect_bundle_type(filename: str) -> Optional[str]:
-    """Detect if file is a SciTeX bundle (.pltz, .figz, .statsz)."""
-    name_lower = filename.lower()
-    for ext in ['.pltz', '.figz', '.statsz']:
-        if name_lower.endswith(ext) or name_lower.endswith(f'{ext}.d'):
-            return ext[1:]  # Return without dot
-    return None
-
-
-def _read_bundle_metadata(bundle_path: Path, bundle_type: str) -> Dict[str, Any]:
-    """Read metadata from a SciTeX bundle."""
-    import zipfile
-
-    result = {
-        "bundle_type": bundle_type,
-        "spec": None,
-        "has_png": False,
-        "has_svg": False,
-        "has_pdf": False,
-        "has_csv": False,
-        "panels": [],
-        "plots": [],
-    }
-
-    # Determine if ZIP or directory
-    is_zip = zipfile.is_zipfile(str(bundle_path))
-
-    if is_zip:
-        with zipfile.ZipFile(bundle_path, 'r') as zf:
-            file_list = zf.namelist()
-
-            # Check for spec file
-            spec_names = {
-                'pltz': 'plot.json',
-                'figz': 'figure.json',
-                'statsz': 'stats.json',
-            }
-            spec_name = spec_names.get(bundle_type)
-
-            if spec_name and spec_name in file_list:
-                with zf.open(spec_name) as f:
-                    result["spec"] = json.load(f)
-
-            # Check for exports
-            result["has_png"] = any(f.endswith('.png') for f in file_list)
-            result["has_svg"] = any(f.endswith('.svg') for f in file_list)
-            result["has_pdf"] = any(f.endswith('.pdf') for f in file_list)
-            result["has_csv"] = any(f.endswith('.csv') for f in file_list)
-
-            # For figz, list panels
-            if bundle_type == 'figz':
-                result["panels"] = [f for f in file_list if '.pltz' in f]
-
-    else:
-        # Directory bundle
-        if bundle_path.is_dir():
-            spec_names = {
-                'pltz': 'plot.json',
-                'figz': 'figure.json',
-                'statsz': 'stats.json',
-            }
-            spec_name = spec_names.get(bundle_type)
-
-            if spec_name:
-                spec_path = bundle_path / spec_name
-                if spec_path.exists():
-                    with open(spec_path, 'r') as f:
-                        result["spec"] = json.load(f)
-
-            # Check for exports
-            result["has_png"] = (bundle_path / "plot.png").exists() or any(bundle_path.glob("*.png"))
-            result["has_svg"] = (bundle_path / "plot.svg").exists() or any(bundle_path.glob("*.svg"))
-            result["has_pdf"] = (bundle_path / "plot.pdf").exists() or any(bundle_path.glob("*.pdf"))
-            result["has_csv"] = (bundle_path / "plot.csv").exists() or any(bundle_path.glob("*.csv"))
-
-            # For figz, list panel directories
-            if bundle_type == 'figz':
-                result["panels"] = [d.name for d in bundle_path.iterdir() if '.pltz' in d.name]
-
-    return result
-
-
-def _get_svg_dimensions(svg_path: str) -> Dict[str, Any]:
-    """Extract dimensions from SVG file."""
-    import re
-
-    try:
-        with open(svg_path, 'r', encoding='utf-8') as f:
-            content = f.read(2000)  # Only read beginning
-
-        # Try to extract viewBox
-        viewbox_match = re.search(r'viewBox=["\']([^"\']+)["\']', content)
-        if viewbox_match:
-            parts = viewbox_match.group(1).split()
-            if len(parts) >= 4:
-                return {
-                    "width": float(parts[2]),
-                    "height": float(parts[3]),
-                    "unit": "viewBox",
-                }
-
-        # Try to extract width/height attributes
-        width_match = re.search(r'width=["\']([0-9.]+)(px|pt|mm|in)?["\']', content)
-        height_match = re.search(r'height=["\']([0-9.]+)(px|pt|mm|in)?["\']', content)
-
-        if width_match and height_match:
-            return {
-                "width": float(width_match.group(1)),
-                "height": float(height_match.group(1)),
-                "unit": width_match.group(2) or "px",
-            }
-
-    except Exception as e:
-        logger.warning(f"Failed to parse SVG dimensions: {e}")
-
-    return {"width": None, "height": None, "unit": None}
+# Re-export for backward compatibility
+__all__ = ["read_image_metadata", "docx2tex_convert"]
 
 
 @csrf_exempt
@@ -154,17 +47,8 @@ def read_image_metadata(request):
     - Images: PNG, JPEG, SVG, WEBP, GIF, TIFF, BMP
     - Documents: PDF
     - SciTeX Bundles: .pltz, .pltz.d, .figz, .figz.d, .statsz, .statsz.d
-
-    Extracts:
-    - Embedded metadata (PNG tEXt, JPEG EXIF, PDF Subject, Bundle spec)
-    - File dimensions (pixels/points/viewBox)
-    - Bundle contents (panels, exports, data)
-
-    Returns:
-        JSON response with metadata, dimensions, and file info
     """
     try:
-        # Check if file was uploaded
         if "image" not in request.FILES:
             return JsonResponse(
                 {"error": "No file provided", "has_metadata": False}, status=400
@@ -172,22 +56,11 @@ def read_image_metadata(request):
 
         uploaded_file = request.FILES["image"]
         filename = uploaded_file.name
-        file_ext = filename.split('.')[-1].lower()
+        file_ext = filename.split(".")[-1].lower()
+        bundle_type = detect_bundle_type(filename)
 
-        # Check if this is a bundle (could be ZIP or directory name)
-        bundle_type = _detect_bundle_type(filename)
-
-        # Save to temporary file/directory
-        if bundle_type:
-            # For bundles, preserve the full extension
-            suffix = ""
-            for ext in ['.pltz.d', '.figz.d', '.statsz.d', '.pltz', '.figz', '.statsz']:
-                if filename.lower().endswith(ext):
-                    suffix = ext
-                    break
-            tmp_file = tempfile.NamedTemporaryFile(delete=False, suffix=suffix)
-        else:
-            tmp_file = tempfile.NamedTemporaryFile(delete=False, suffix=f".{file_ext}")
+        suffix = _get_temp_suffix(filename, bundle_type, file_ext)
+        tmp_file = tempfile.NamedTemporaryFile(delete=False, suffix=suffix)
 
         for chunk in uploaded_file.chunks():
             tmp_file.write(chunk)
@@ -195,161 +68,19 @@ def read_image_metadata(request):
         tmp_path = Path(tmp_file.name)
 
         try:
-            # Import scitex.io
             from scitex.io import load, read_metadata
 
-            # Handle SciTeX bundles
             if bundle_type:
-                bundle_info = _read_bundle_metadata(tmp_path, bundle_type)
-
-                response_data = {
-                    "has_metadata": bundle_info["spec"] is not None,
-                    "metadata": bundle_info["spec"],
-                    "file_type": f"bundle_{bundle_type}",
-                    "bundle_info": {
-                        "type": bundle_type,
-                        "has_png": bundle_info["has_png"],
-                        "has_svg": bundle_info["has_svg"],
-                        "has_pdf": bundle_info["has_pdf"],
-                        "has_csv": bundle_info["has_csv"],
-                        "panels": bundle_info.get("panels", []),
-                    },
-                    "message": f"SciTeX {bundle_type.upper()} bundle loaded",
-                }
-
-                # Try to get dimensions from PNG if available
-                if bundle_info["has_png"]:
-                    try:
-                        import zipfile
-                        if zipfile.is_zipfile(str(tmp_path)):
-                            with zipfile.ZipFile(tmp_path, 'r') as zf:
-                                png_files = [f for f in zf.namelist() if f.endswith('.png')]
-                                if png_files:
-                                    with zf.open(png_files[0]) as png_f:
-                                        from PIL import Image
-                                        import io
-                                        img = Image.open(io.BytesIO(png_f.read()))
-                                        response_data["dimensions"] = {
-                                            "width": img.width,
-                                            "height": img.height,
-                                        }
-                                        img.close()
-                        elif tmp_path.is_dir():
-                            png_files = list(tmp_path.glob("*.png"))
-                            if png_files:
-                                from PIL import Image
-                                img = Image.open(png_files[0])
-                                response_data["dimensions"] = {
-                                    "width": img.width,
-                                    "height": img.height,
-                                }
-                                img.close()
-                    except Exception as e:
-                        logger.warning(f"Could not read bundle PNG dimensions: {e}")
-
-                return JsonResponse(response_data)
-
-            # Handle SVG files
-            elif file_ext == 'svg':
-                svg_dims = _get_svg_dimensions(str(tmp_path))
-
-                # Try to extract metadata from SVG (look for scitex metadata comment)
-                metadata = None
-                try:
-                    with open(tmp_path, 'r', encoding='utf-8') as f:
-                        content = f.read()
-                        # Look for embedded JSON metadata in comment
-                        import re
-                        meta_match = re.search(r'<!--\s*scitex_metadata:\s*(\{.*?\})\s*-->', content, re.DOTALL)
-                        if meta_match:
-                            metadata = json.loads(meta_match.group(1))
-                except Exception as e:
-                    logger.warning(f"Failed to extract SVG metadata: {e}")
-
-                response_data = {
-                    "has_metadata": metadata is not None,
-                    "metadata": metadata,
-                    "file_type": "svg",
-                    "dimensions": {
-                        "width": svg_dims.get("width"),
-                        "height": svg_dims.get("height"),
-                        "unit": svg_dims.get("unit", "px"),
-                    },
-                    "message": "SVG file loaded (vector format)",
-                }
-
-                return JsonResponse(response_data)
-
-            # Handle PDF files
-            elif file_ext == 'pdf':
-                # Extract metadata using scitex.io
-                metadata = read_metadata(str(tmp_path))
-
-                # For PDFs, load with metadata mode to get page info
-                pdf_data = load(str(tmp_path), mode='metadata')
-
-                response_data = {
-                    "has_metadata": metadata is not None,
-                    "metadata": metadata,
-                    "file_type": "pdf",
-                    "page_count": pdf_data.get('pages', 0),
-                    "dimensions": {
-                        "width_pt": None,  # Need first page for this
-                        "height_pt": None,
-                        "pages": pdf_data.get('pages', 0),
-                    },
-                    "pdf_metadata": {
-                        "title": pdf_data.get('title', ''),
-                        "author": pdf_data.get('author', ''),
-                        "subject": pdf_data.get('subject', ''),
-                        "creator": pdf_data.get('creator', ''),
-                    }
-                }
-
-                if metadata is None:
-                    response_data["message"] = "No SciTeX metadata found in PDF"
-                else:
-                    response_data["message"] = "Metadata successfully extracted"
-
-                return JsonResponse(response_data)
-
-            # Handle raster images
+                return _handle_bundle(tmp_path, bundle_type)
+            elif file_ext == "svg":
+                return _handle_svg(tmp_path)
+            elif file_ext == "pdf":
+                return _handle_pdf(tmp_path, read_metadata, load)
             else:
-                # Extract metadata using scitex.io
-                metadata = read_metadata(str(tmp_path))
-
-                # Load to get dimensions
-                img, _ = load(str(tmp_path), metadata=True)
-
-                response_data = {
-                    "has_metadata": metadata is not None,
-                    "metadata": metadata,
-                    "file_type": "image",
-                    "dimensions": {
-                        "width": img.width,
-                        "height": img.height,
-                    }
-                }
-
-                img.close()
-
-                if metadata is None:
-                    response_data["message"] = f"No SciTeX metadata found in {file_ext.upper()}"
-                else:
-                    response_data["message"] = "Metadata successfully extracted"
-
-                return JsonResponse(response_data)
+                return _handle_raster_image(tmp_path, file_ext, read_metadata, load)
 
         finally:
-            # Clean up temporary file
-            import os
-            import shutil
-
-            if tmp_path.exists():
-                if tmp_path.is_dir():
-                    shutil.rmtree(tmp_path)
-                else:
-                    os.unlink(tmp_path)
+            _cleanup_temp(tmp_path)
 
     except ImportError as e:
         logger.error(f"Failed to import scitex.io: {e}")
@@ -369,151 +100,133 @@ def read_image_metadata(request):
         )
 
 
-@csrf_exempt
-@require_http_methods(["POST"])
-def docx2tex_convert(request):
-    """
-    Convert DOCX file to LaTeX using scitex.msword.
+def _get_temp_suffix(filename: str, bundle_type, file_ext: str) -> str:
+    """Get appropriate temp file suffix."""
+    if bundle_type:
+        for ext in [".pltz.d", ".figz.d", ".statsz.d", ".pltz", ".figz", ".statsz"]:
+            if filename.lower().endswith(ext):
+                return ext
+    return f".{file_ext}"
 
-    Accepts:
-    - file: The uploaded .docx file
-    - profile: Journal profile name (default: "generic")
-    - link_mode: "by-number" or "by-proximity" (default: "by-number")
-    - normalize_headings: bool (default: True)
-    - validate: bool (default: True)
-    - extract_images: bool (default: True)
 
-    Returns:
-        JSON response with:
-        - latex: The generated LaTeX content
-        - metadata: Document metadata
-        - blocks: Document structure blocks
-        - images: Extracted images (base64 encoded if extract_images=True)
-        - references: Parsed references
-        - warnings: Document validation warnings
-    """
+def _handle_bundle(tmp_path: Path, bundle_type: str) -> JsonResponse:
+    """Handle SciTeX bundle files."""
+    bundle_info = read_bundle_metadata(tmp_path, bundle_type)
+
+    response_data = {
+        "has_metadata": bundle_info["spec"] is not None,
+        "metadata": bundle_info["spec"],
+        "file_type": f"bundle_{bundle_type}",
+        "bundle_info": {
+            "type": bundle_type,
+            "has_png": bundle_info["has_png"],
+            "has_svg": bundle_info["has_svg"],
+            "has_pdf": bundle_info["has_pdf"],
+            "has_csv": bundle_info["has_csv"],
+            "panels": bundle_info.get("panels", []),
+        },
+        "message": f"SciTeX {bundle_type.upper()} bundle loaded",
+    }
+
+    if bundle_info["has_png"]:
+        dims = get_bundle_dimensions_from_png(tmp_path)
+        if dims:
+            response_data["dimensions"] = dims
+
+    return JsonResponse(response_data)
+
+
+def _handle_svg(tmp_path: Path) -> JsonResponse:
+    """Handle SVG files."""
+    svg_dims = get_svg_dimensions(str(tmp_path))
+    metadata = None
+
     try:
-        # Check if file was uploaded
-        if "file" not in request.FILES:
-            return JsonResponse({"error": "No file provided"}, status=400)
-
-        uploaded_file = request.FILES["file"]
-
-        # Validate file extension
-        if not uploaded_file.name.lower().endswith(".docx"):
-            return JsonResponse(
-                {"error": "Invalid file type. Please upload a .docx file"},
-                status=400,
+        with open(tmp_path, "r", encoding="utf-8") as f:
+            content = f.read()
+            meta_match = re.search(
+                r"<!--\s*scitex_metadata:\s*(\{.*?\})\s*-->", content, re.DOTALL
             )
-
-        # Parse options from form data
-        profile = request.POST.get("profile", "generic")
-        link_mode = request.POST.get("link_mode", "by-number")
-        normalize_headings = request.POST.get("normalize_headings", "true").lower() == "true"
-        validate = request.POST.get("validate", "true").lower() == "true"
-        extract_images = request.POST.get("extract_images", "true").lower() == "true"
-
-        # Save to temporary file
-        with tempfile.NamedTemporaryFile(
-            delete=False, suffix=".docx"
-        ) as tmp_file:
-            for chunk in uploaded_file.chunks():
-                tmp_file.write(chunk)
-            tmp_path = Path(tmp_file.name)
-
-        try:
-            # Import scitex.msword
-            from scitex.msword import (
-                load_docx,
-                normalize_section_headings,
-                link_captions_to_images,
-                link_captions_to_images_by_proximity,
-                validate_document,
-            )
-            from scitex.tex import export_tex
-
-            # 1. Load DOCX
-            doc = load_docx(tmp_path, profile=profile, extract_images=extract_images)
-
-            # 2. Normalize headings (optional)
-            if normalize_headings:
-                doc = normalize_section_headings(doc)
-
-            # 3. Link captions to images (optional)
-            if extract_images and doc.get("images"):
-                if link_mode == "by-proximity":
-                    doc = link_captions_to_images_by_proximity(doc)
-                else:
-                    doc = link_captions_to_images(doc)
-
-            # 4. Validate document (optional)
-            if validate:
-                doc = validate_document(doc)
-
-            # 5. Export to LaTeX
-            with tempfile.NamedTemporaryFile(
-                delete=False, suffix=".tex", mode="w"
-            ) as tex_file:
-                tex_path = Path(tex_file.name)
-
-            # Create temp directory for images
-            with tempfile.TemporaryDirectory() as img_dir:
-                tex_path = export_tex(doc, tex_path, image_dir=img_dir)
-
-                # Read the generated LaTeX
-                latex_content = tex_path.read_text()
-
-            # Prepare images for JSON response (base64 encode)
-            images_response = []
-            for img in doc.get("images", []):
-                img_data = {
-                    "rel_id": img.get("rel_id", ""),
-                    "hash": img.get("hash", ""),
-                    "content_type": img.get("content_type", "image/png"),
-                    "extension": img.get("extension", ".png"),
-                    "size_bytes": img.get("size_bytes", 0),
-                }
-                # Include base64 data if available
-                if "data" in img and img["data"]:
-                    img_data["data"] = base64.b64encode(img["data"]).decode("utf-8")
-                images_response.append(img_data)
-
-            # Build response
-            response_data = {
-                "latex": latex_content,
-                "metadata": doc.get("metadata", {}),
-                "blocks": doc.get("blocks", []),
-                "images": images_response,
-                "references": doc.get("references", []),
-                "warnings": doc.get("warnings", []),
-            }
-
-            return JsonResponse(response_data)
-
-        finally:
-            # Clean up temporary files
-            import os
-
-            if tmp_path.exists():
-                os.unlink(tmp_path)
-            if tex_path.exists():
-                os.unlink(tex_path)
-
-    except ImportError as e:
-        logger.error(f"Failed to import scitex.msword: {e}")
-        return JsonResponse(
-            {"error": "DOCX conversion not available (scitex.msword not installed)"},
-            status=500,
-        )
-
+            if meta_match:
+                metadata = json.loads(meta_match.group(1))
     except Exception as e:
-        logger.error(f"Error converting DOCX to LaTeX: {e}")
-        import traceback
-        traceback.print_exc()
-        return JsonResponse(
-            {"error": f"Conversion failed: {str(e)}"},
-            status=500,
-        )
+        logger.warning(f"Failed to extract SVG metadata: {e}")
+
+    return JsonResponse(
+        {
+            "has_metadata": metadata is not None,
+            "metadata": metadata,
+            "file_type": "svg",
+            "dimensions": {
+                "width": svg_dims.get("width"),
+                "height": svg_dims.get("height"),
+                "unit": svg_dims.get("unit", "px"),
+            },
+            "message": "SVG file loaded (vector format)",
+        }
+    )
+
+
+def _handle_pdf(tmp_path: Path, read_metadata, load) -> JsonResponse:
+    """Handle PDF files."""
+    metadata = read_metadata(str(tmp_path))
+    pdf_data = load(str(tmp_path), mode="metadata")
+
+    response_data = {
+        "has_metadata": metadata is not None,
+        "metadata": metadata,
+        "file_type": "pdf",
+        "page_count": pdf_data.get("pages", 0),
+        "dimensions": {
+            "width_pt": None,
+            "height_pt": None,
+            "pages": pdf_data.get("pages", 0),
+        },
+        "pdf_metadata": {
+            "title": pdf_data.get("title", ""),
+            "author": pdf_data.get("author", ""),
+            "subject": pdf_data.get("subject", ""),
+            "creator": pdf_data.get("creator", ""),
+        },
+    }
+
+    response_data["message"] = (
+        "Metadata successfully extracted"
+        if metadata
+        else "No SciTeX metadata found in PDF"
+    )
+    return JsonResponse(response_data)
+
+
+def _handle_raster_image(tmp_path: Path, file_ext: str, read_metadata, load):
+    """Handle raster image files."""
+    metadata = read_metadata(str(tmp_path))
+    img, _ = load(str(tmp_path), metadata=True)
+
+    response_data = {
+        "has_metadata": metadata is not None,
+        "metadata": metadata,
+        "file_type": "image",
+        "dimensions": {"width": img.width, "height": img.height},
+    }
+
+    img.close()
+
+    response_data["message"] = (
+        "Metadata successfully extracted"
+        if metadata
+        else f"No SciTeX metadata found in {file_ext.upper()}"
+    )
+    return JsonResponse(response_data)
+
+
+def _cleanup_temp(tmp_path: Path):
+    """Clean up temporary file or directory."""
+    if tmp_path.exists():
+        if tmp_path.is_dir():
+            shutil.rmtree(tmp_path)
+        else:
+            os.unlink(tmp_path)
 
 
 # EOF
