+++ apps/public_app/views/tools.py
@@ -1,479 +1,91 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-# Timestamp: "2025-11-28 21:31:00 (ywatanabe)"
-# File: /home/ywatanabe/proj/scitex-cloud/apps/public_app/views/tools.py
-# ----------------------------------------
+"""Research Tools views - Re-export wrapper.
+
+This module re-exports all tool views and data for backward compatibility.
+Actual implementation is in tools_views.py and tools_data.py.
+"""
+
 from __future__ import annotations
-import os
 
-__FILE__ = "./apps/public_app/views/tools.py"
-__DIR__ = os.path.dirname(__FILE__)
-# ----------------------------------------
-
-"""
-Research Tools Views
-
-Handles all research tool pages and utilities.
-Organized by tool type: Text, Image, PDF, Video, Rendering, Developer, Research
-"""
-
-from django.shortcuts import render
-
-
-def tools(request):
-    """Research tools page - bookmarklets and utilities for researchers."""
-    # Organized by tool type: Text → Image → PDF → Video → Rendering → Developer → Research
-    domains = [
-        {
-            "name": "Text",
-            "slug": "text",
-            "icon": "📝",
-            "description": "Format, compare, and process text content",
-            "tools": [
-                {
-                    "name": "Markdown Renderer",
-                    "description": "Real-time Markdown preview with syntax highlighting and table support.",
-                    "use_case": "Format README files and documentation for data repositories",
-                    "bookmarklet_url": "/tools/markdown-renderer/",
-                    "icon": "📝",
-                },
-                {
-                    "name": "Text Diff Checker",
-                    "description": "Compare two text blocks side-by-side with highlighted differences.",
-                    "use_case": "Compare dataset versions or track changes in results",
-                    "bookmarklet_url": "/tools/text-diff-checker/",
-                    "icon": "🔄",
-                },
-                {
-                    "name": "JSON Formatter",
-                    "description": "Format, validate, and beautify JSON data with syntax highlighting.",
-                    "use_case": "Validate plot specifications and configuration files",
-                    "bookmarklet_url": "/tools/json-formatter/",
-                    "icon": "{ }",
-                },
-            ],
-        },
-        {
-            "name": "Image",
-            "slug": "image",
-            "icon": "🖼️",
-            "description": "Manipulate and convert images for publications",
-            "tools": [
-                {
-                    "name": "Image & PDF Viewer",
-                    "description": "View dimensions, DPI, and unit conversions (mm/inch) for publication figures.",
-                    "use_case": "Verify Figure 2 meets journal dimension requirements",
-                    "bookmarklet_url": "/tools/image-viewer/",
-                    "icon": "📐",
-                },
-                {
-                    "name": "Image Resizer",
-                    "description": "Resize and crop images for journal submissions with preset dimensions.",
-                    "use_case": "Adjust Figure 2 to exact pixel-perfect journal specs",
-                    "bookmarklet_url": "/tools/image-resizer/",
-                    "icon": "📏",
-                },
-                {
-                    "name": "Image Converter",
-                    "description": "Convert images between PNG, JPG, WEBP, TIFF formats with batch conversion.",
-                    "use_case": "Convert PNG figures to TIFF for journal submission",
-                    "bookmarklet_url": "/tools/image-converter/",
-                    "icon": "🔄",
-                },
-                {
-                    "name": "Image Concatenator",
-                    "description": "Combine multiple images into a single tiled multi-panel figure.",
-                    "use_case": "Create Figure 1 panel layouts (A, B, C, D)",
-                    "bookmarklet_url": "/tools/image-concatenator/",
-                    "icon": "🖼️",
-                },
-                {
-                    "name": "Mermaid Diagram Renderer",
-                    "description": "Create flowcharts, sequence diagrams, and concept diagrams from text syntax.",
-                    "use_case": "Design experimental workflow diagrams for Methods section",
-                    "bookmarklet_url": "/tools/mermaid-renderer/",
-                    "icon": "🧜‍♀️",
-                },
-                {
-                    "name": "Images to GIF",
-                    "description": "Convert image sequences into animated GIF with customizable duration.",
-                    "use_case": "Create supplementary animations showing temporal changes",
-                    "bookmarklet_url": "/tools/images-to-gif/",
-                    "icon": "🎬",
-                },
-                {
-                    "name": "Images to PDF",
-                    "description": "Convert multiple images into a single PDF with custom page orientation.",
-                    "use_case": "Create supplementary figures PDF from multiple images",
-                    "bookmarklet_url": "/tools/images-to-pdf/",
-                    "icon": "📄",
-                },
-                {
-                    "name": "PDF to Images",
-                    "description": "Extract all pages from PDF as PNG or JPG images with adjustable DPI.",
-                    "use_case": "Convert PDF figures to images for presentation slides",
-                    "bookmarklet_url": "/tools/pdf-to-images/",
-                    "icon": "🖼️",
-                },
-            ],
-        },
-        {
-            "name": "PDF",
-            "slug": "pdf",
-            "icon": "📄",
-            "description": "Manage and process PDF documents",
-            "tools": [
-                {
-                    "name": "Image & PDF Viewer",
-                    "description": "View dimensions, DPI, and unit conversions (mm/inch) for publication figures.",
-                    "use_case": "Verify Figure 2 meets journal dimension requirements",
-                    "bookmarklet_url": "/tools/image-viewer/",
-                    "icon": "📐",
-                },
-                {
-                    "name": "PDF Merger",
-                    "description": "Combine multiple PDF files into a single document with drag-to-reorder.",
-                    "use_case": "Merge manuscript, figures, and supplements for submission",
-                    "bookmarklet_url": "/tools/pdf-merger/",
-                    "icon": "📑",
-                },
-                {
-                    "name": "PDF Compressor",
-                    "description": "Reduce PDF file size while maintaining quality for email and uploads.",
-                    "use_case": "Compress submission files under journal size limits",
-                    "bookmarklet_url": "/tools/pdf-compressor/",
-                    "icon": "🗜️",
-                },
-                {
-                    "name": "PDF Splitter",
-                    "description": "Extract specific pages from PDF files using page ranges.",
-                    "use_case": "Extract figures from compiled manuscript for separate upload",
-                    "bookmarklet_url": "/tools/pdf-splitter/",
-                    "icon": "✂️",
-                },
-                {
-                    "name": "Images to PDF",
-                    "description": "Convert multiple images into a single PDF with custom page orientation.",
-                    "use_case": "Create supplementary figures PDF from multiple images",
-                    "bookmarklet_url": "/tools/images-to-pdf/",
-                    "icon": "📄",
-                },
-                {
-                    "name": "PDF to Images",
-                    "description": "Extract all pages from PDF as PNG or JPG images with adjustable DPI.",
-                    "use_case": "Convert PDF figures to images for presentation slides",
-                    "bookmarklet_url": "/tools/pdf-to-images/",
-                    "icon": "🖼️",
-                },
-            ],
-        },
-        {
-            "name": "Video",
-            "slug": "video",
-            "icon": "🎬",
-            "description": "Video and animation processing",
-            "tools": [
-                {
-                    "name": "Video Editor",
-                    "description": "Trim videos by time window with browser-based processing.",
-                    "use_case": "Edit supplementary videos for journal submission",
-                    "bookmarklet_url": "/tools/video-editor/",
-                    "icon": "🎬",
-                },
-                {
-                    "name": "Images to GIF",
-                    "description": "Convert image sequences into animated GIF with customizable duration.",
-                    "use_case": "Create supplementary animations showing temporal changes",
-                    "bookmarklet_url": "/tools/images-to-gif/",
-                    "icon": "🎬",
-                },
-            ],
-        },
-        {
-            "name": "Rendering",
-            "slug": "rendering",
-            "icon": "📈",
-            "description": "Create publication-quality plots and diagrams",
-            "tools": [
-                {
-                    "name": "Plot Viewer",
-                    "description": "Interactive CSV plot viewer with Nature journal standards. Supports line, scatter, and bar plots with 300 DPI rendering.",
-                    "use_case": "Quick data visualization during analysis",
-                    "bookmarklet_url": "/tools/plot-viewer/",
-                    "icon": "📊",
-                },
-                {
-                    "name": "Plot Backend Test",
-                    "description": "Test matplotlib/scitex.plt backend with JSON specifications. Generate publication-quality SVG plots.",
-                    "use_case": "Design figures with precise journal specifications",
-                    "bookmarklet_url": "/tools/plot-backend-test/",
-                    "icon": "🧪",
-                },
-                {
-                    "name": "Mermaid Diagram Renderer",
-                    "description": "Create flowcharts, sequence diagrams, and concept diagrams from text syntax.",
-                    "use_case": "Design experimental workflow diagrams for Methods section",
-                    "bookmarklet_url": "/tools/mermaid-renderer/",
-                    "icon": "🧜‍♀️",
-                },
-                {
-                    "name": "Markdown Renderer",
-                    "description": "Real-time Markdown preview with syntax highlighting and table support.",
-                    "use_case": "Format README files and documentation for data repositories",
-                    "bookmarklet_url": "/tools/markdown-renderer/",
-                    "icon": "📝",
-                },
-                {
-                    "name": "Color Picker",
-                    "description": "Advanced color picker with format conversion and palette generation.",
-                    "use_case": "Design consistent color schemes for figure panels",
-                    "bookmarklet_url": "/tools/color-picker/",
-                    "icon": "🎨",
-                },
-            ],
-        },
-        {
-            "name": "Developer",
-            "slug": "development",
-            "icon": "💻",
-            "description": "Web development and debugging utilities",
-            "tools": [
-                {
-                    "name": "Element Inspector",
-                    "description": "Visual debugging tool with AI-ready output format.",
-                    "use_case": "Debug web interface issues in research platforms",
-                    "bookmarklet_url": "/tools/element-inspector/",
-                    "icon": "🔍",
-                },
-                {
-                    "name": "Repository Concatenator",
-                    "description": "Concatenate repository files into AI-ready format for code review.",
-                    "use_case": "Prepare analysis scripts for AI code review",
-                    "bookmarklet_url": "/tools/repo-concatenator/",
-                    "icon": "📦",
-                },
-                {
-                    "name": "Color Picker",
-                    "description": "Advanced color picker with format conversion and palette generation.",
-                    "use_case": "Design consistent color schemes for figure panels",
-                    "bookmarklet_url": "/tools/color-picker/",
-                    "icon": "🎨",
-                },
-                {
-                    "name": "QR Code Generator",
-                    "description": "Generate QR codes for URLs, DOIs, posters, and presentations.",
-                    "use_case": "Add QR codes to conference posters linking to papers",
-                    "bookmarklet_url": "/tools/qr-code-generator/",
-                    "icon": "📱",
-                },
-            ],
-        },
-        {
-            "name": "Research",
-            "slug": "research",
-            "icon": "🔬",
-            "description": "Literature management and citation tools",
-            "tools": [
-                {
-                    "name": "Asta AI Citation Scraper",
-                    "description": "Automatically collect all BibTeX citations from Asta AI search results.",
-                    "use_case": "Build bibliography from AI literature searches",
-                    "bookmarklet_url": "/tools/asta-citation-scraper/",
-                    "icon": "📚",
-                },
-                {
-                    "name": "Statistics Calculator",
-                    "description": "Quick statistical analysis for research data with descriptive stats, t-tests, and correlations.",
-                    "use_case": "Verify experimental results before plotting",
-                    "bookmarklet_url": "/tools/statistics-calculator/",
-                    "icon": "📈",
-                },
-            ],
-        },
-        {
-            "name": "Document Conversion",
-            "slug": "document",
-            "icon": "📄",
-            "description": "Convert documents between formats",
-            "tools": [
-                {
-                    "name": "DOCX to LaTeX",
-                    "description": "Convert Microsoft Word documents to LaTeX format with figure extraction and reference parsing.",
-                    "use_case": "Convert Word manuscript to LaTeX for journal submission",
-                    "bookmarklet_url": "/tools/docx2tex/",
-                    "icon": "📝",
-                },
-            ],
-        },
-    ]
-
-    # Calculate total tools
-    total_tools = sum(len(domain["tools"]) for domain in domains)
-
-    context = {
-        "domains": domains,
-        "total_tools": total_tools,
-    }
-
-    return render(request, "public_app/pages/tools.html", context)
-
-
-# Individual tool detail pages
-def tool_element_inspector(request):
-    """Element Inspector tool detail page."""
-    return render(request, "public_app/tools/element-inspector.html")
-
-
-def tool_asta_citation_scraper(request):
-    """Asta AI Citation Scraper tool detail page."""
-    return render(request, "public_app/tools/asta-citation-scraper.html")
-
-
-def tool_image_concatenator(request):
-    """Image Concatenator tool detail page."""
-    return render(request, "public_app/tools/image-concatenator.html")
-
-
-def tool_qr_code_generator(request):
-    """QR Code Generator tool detail page."""
-    return render(request, "public_app/tools/qr-code-generator.html")
-
-
-def tool_color_picker(request):
-    """Color Picker tool detail page."""
-    return render(request, "public_app/tools/color-picker.html")
-
-
-def tool_markdown_renderer(request):
-    """Markdown Renderer tool detail page."""
-    return render(request, "public_app/tools/markdown-renderer.html")
-
-
-def tool_text_diff_checker(request):
-    """Text Diff Checker tool detail page."""
-    return render(request, "public_app/tools/text-diff-checker.html")
-
-
-def tool_images_to_gif(request):
-    """Images to GIF tool detail page."""
-    return render(request, "public_app/tools/images-to-gif.html")
-
-
-def tool_image_converter(request):
-    """Image Converter tool detail page."""
-    return render(request, "public_app/tools/image-converter.html")
-
-
-def tool_pdf_merger(request):
-    """PDF Merger tool detail page."""
-    return render(request, "public_app/tools/pdf-merger.html")
-
-
-def tool_statistics_calculator(request):
-    """Statistics Calculator tool detail page."""
-    return render(request, "public_app/tools/statistics-calculator.html")
-
-
-def tool_pdf_splitter(request):
-    """PDF Splitter tool detail page."""
-    return render(request, "public_app/tools/pdf-splitter.html")
-
-
-def tool_image_resizer(request):
-    """Image Resizer tool detail page."""
-    return render(request, "public_app/tools/image-resizer.html")
-
-
-def tool_repo_concatenator(request):
-    """Repository Concatenator tool detail page."""
-    return render(request, "public_app/tools/repo-concatenator.html")
-
-
-def tool_json_formatter(request):
-    """JSON Formatter tool detail page."""
-    return render(request, "public_app/tools/json-formatter.html")
-
-
-def tool_images_to_pdf(request):
-    """Images to PDF tool detail page."""
-    return render(request, "public_app/tools/images-to-pdf.html")
-
-
-def tool_pdf_to_images(request):
-    """PDF to Images tool detail page."""
-    return render(request, "public_app/tools/pdf-to-images.html")
-
-
-def tool_pdf_compressor(request):
-    """PDF Compressor tool detail page."""
-    return render(request, "public_app/tools/pdf-compressor.html")
-
-
-def tool_video_editor(request):
-    """Video Editor tool detail page."""
-    return render(request, "public_app/tools/video-editor.html")
-
-
-def tool_plot_viewer(request):
-    """
-    Quick CSV Plot Viewer - renders simple CSV plots using Canvas.
-
-    Accepts CSV files with column naming convention:
-    ax_{axis_index}_{plot_id}_{plot_type}_{variable}
-
-    Example: ax_00_plot_line_0_line_x, ax_00_plot_line_0_line_y
-
-    Supports: line, scatter, bar plots only.
-    For advanced plot types, use the backend plot API.
-    """
-    return render(request, "public_app/tools/plot-viewer.html")
-
-
-def tool_plot_backend_test(request):
-    """
-    Backend Plot Renderer Test - test matplotlib/scitex.plt backend.
-
-    Internal testing tool for the backend plot API.
-    """
-    return render(request, "public_app/tools/plot-backend-test.html")
-
-
-def tool_image_viewer(request):
-    """
-    Image Viewer - Display image with dimension, DPI, and unit conversion info.
-
-    Shows pixel dimensions, DPI, physical size (mm/inch), and conversions
-    to help understand figure dimensions for publications.
-    """
-    return render(request, "public_app/tools/image-viewer.html")
-
-
-def tool_mermaid_renderer(request):
-    """
-    Mermaid Diagram Renderer - Create diagrams from text syntax.
-
-    Supports flowcharts, sequence diagrams, Gantt charts, class diagrams,
-    pie charts, and git graphs using Mermaid.js syntax.
-    """
-    return render(request, "public_app/tools/mermaid-renderer.html")
-
-
-def tool_docx2tex(request):
-    """
-    DOCX to LaTeX Converter - Convert Word documents to LaTeX.
-
-    Uses scitex.msword to convert DOCX files to LaTeX format with:
-    - Figure extraction and linking
-    - Reference parsing
-    - Section heading normalization
-    - Multiple journal profile support
-    """
-    from scitex.msword import list_profiles
-
-    profiles = list_profiles()
-    return render(
-        request, "public_app/tools/docx2tex.html", {"profiles": profiles}
-    )
+from .tools_data import (
+    DEVELOPER_TOOLS,
+    DOCUMENT_TOOLS,
+    IMAGE_TOOLS,
+    PDF_TOOLS,
+    RENDERING_TOOLS,
+    RESEARCH_TOOLS,
+    TEXT_TOOLS,
+    VIDEO_TOOLS,
+    get_tool_domains,
+)
+from .tools_views import (
+    tool_asta_citation_scraper,
+    tool_color_picker,
+    tool_docx2tex,
+    tool_element_inspector,
+    tool_image_concatenator,
+    tool_image_converter,
+    tool_image_resizer,
+    tool_image_viewer,
+    tool_images_to_gif,
+    tool_images_to_pdf,
+    tool_json_formatter,
+    tool_markdown_renderer,
+    tool_mermaid_renderer,
+    tool_pdf_compressor,
+    tool_pdf_merger,
+    tool_pdf_splitter,
+    tool_pdf_to_images,
+    tool_plot_backend_test,
+    tool_plot_viewer,
+    tool_qr_code_generator,
+    tool_repo_concatenator,
+    tool_statistics_calculator,
+    tool_text_diff_checker,
+    tool_video_editor,
+    tools,
+)
+
+__all__ = [
+    # Main tools page
+    "tools",
+    # Tool data
+    "get_tool_domains",
+    "TEXT_TOOLS",
+    "IMAGE_TOOLS",
+    "PDF_TOOLS",
+    "VIDEO_TOOLS",
+    "RENDERING_TOOLS",
+    "DEVELOPER_TOOLS",
+    "RESEARCH_TOOLS",
+    "DOCUMENT_TOOLS",
+    # Individual tool views
+    "tool_asta_citation_scraper",
+    "tool_color_picker",
+    "tool_docx2tex",
+    "tool_element_inspector",
+    "tool_image_concatenator",
+    "tool_image_converter",
+    "tool_image_resizer",
+    "tool_image_viewer",
+    "tool_images_to_gif",
+    "tool_images_to_pdf",
+    "tool_json_formatter",
+    "tool_markdown_renderer",
+    "tool_mermaid_renderer",
+    "tool_pdf_compressor",
+    "tool_pdf_merger",
+    "tool_pdf_splitter",
+    "tool_pdf_to_images",
+    "tool_plot_backend_test",
+    "tool_plot_viewer",
+    "tool_qr_code_generator",
+    "tool_repo_concatenator",
+    "tool_statistics_calculator",
+    "tool_text_diff_checker",
+    "tool_video_editor",
+]
 
 
 # EOF
