+++ apps/public_app/views/tools_data_other.py
@@ -1,0 +1,126 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""Research Tools data - Video, Rendering, Developer, Research, Document domains."""
+
+from __future__ import annotations
+
+# Video tools
+VIDEO_TOOLS = [
+    {
+        "name": "Video Editor",
+        "description": "Trim videos by time window with browser-based processing.",
+        "use_case": "Edit supplementary videos for journal submission",
+        "bookmarklet_url": "/tools/video-editor/",
+        "icon": "🎬",
+    },
+    {
+        "name": "Images to GIF",
+        "description": "Convert image sequences into animated GIF with customizable duration.",
+        "use_case": "Create supplementary animations showing temporal changes",
+        "bookmarklet_url": "/tools/images-to-gif/",
+        "icon": "🎬",
+    },
+]
+
+# Rendering tools
+RENDERING_TOOLS = [
+    {
+        "name": "Plot Viewer",
+        "description": "Interactive CSV plot viewer with Nature journal standards.",
+        "use_case": "Quick data visualization during analysis",
+        "bookmarklet_url": "/tools/plot-viewer/",
+        "icon": "📊",
+    },
+    {
+        "name": "Plot Backend Test",
+        "description": "Test matplotlib/scitex.plt backend with JSON specifications.",
+        "use_case": "Design figures with precise journal specifications",
+        "bookmarklet_url": "/tools/plot-backend-test/",
+        "icon": "🧪",
+    },
+    {
+        "name": "Mermaid Diagram Renderer",
+        "description": "Create flowcharts, sequence diagrams, and concept diagrams from text syntax.",
+        "use_case": "Design experimental workflow diagrams for Methods section",
+        "bookmarklet_url": "/tools/mermaid-renderer/",
+        "icon": "🧜‍♀️",
+    },
+    {
+        "name": "Markdown Renderer",
+        "description": "Real-time Markdown preview with syntax highlighting and table support.",
+        "use_case": "Format README files and documentation for data repositories",
+        "bookmarklet_url": "/tools/markdown-renderer/",
+        "icon": "📝",
+    },
+    {
+        "name": "Color Picker",
+        "description": "Advanced color picker with format conversion and palette generation.",
+        "use_case": "Design consistent color schemes for figure panels",
+        "bookmarklet_url": "/tools/color-picker/",
+        "icon": "🎨",
+    },
+]
+
+# Developer tools
+DEVELOPER_TOOLS = [
+    {
+        "name": "Element Inspector",
+        "description": "Visual debugging tool with AI-ready output format.",
+        "use_case": "Debug web interface issues in research platforms",
+        "bookmarklet_url": "/tools/element-inspector/",
+        "icon": "🔍",
+    },
+    {
+        "name": "Repository Concatenator",
+        "description": "Concatenate repository files into AI-ready format for code review.",
+        "use_case": "Prepare analysis scripts for AI code review",
+        "bookmarklet_url": "/tools/repo-concatenator/",
+        "icon": "📦",
+    },
+    {
+        "name": "Color Picker",
+        "description": "Advanced color picker with format conversion and palette generation.",
+        "use_case": "Design consistent color schemes for figure panels",
+        "bookmarklet_url": "/tools/color-picker/",
+        "icon": "🎨",
+    },
+    {
+        "name": "QR Code Generator",
+        "description": "Generate QR codes for URLs, DOIs, posters, and presentations.",
+        "use_case": "Add QR codes to conference posters linking to papers",
+        "bookmarklet_url": "/tools/qr-code-generator/",
+        "icon": "📱",
+    },
+]
+
+# Research tools
+RESEARCH_TOOLS = [
+    {
+        "name": "Asta AI Citation Scraper",
+        "description": "Automatically collect all BibTeX citations from Asta AI search results.",
+        "use_case": "Build bibliography from AI literature searches",
+        "bookmarklet_url": "/tools/asta-citation-scraper/",
+        "icon": "📚",
+    },
+    {
+        "name": "Statistics Calculator",
+        "description": "Quick statistical analysis with descriptive stats, t-tests, and correlations.",
+        "use_case": "Verify experimental results before plotting",
+        "bookmarklet_url": "/tools/statistics-calculator/",
+        "icon": "📈",
+    },
+]
+
+# Document conversion tools
+DOCUMENT_TOOLS = [
+    {
+        "name": "DOCX to LaTeX",
+        "description": "Convert Microsoft Word documents to LaTeX format with figure extraction.",
+        "use_case": "Convert Word manuscript to LaTeX for journal submission",
+        "bookmarklet_url": "/tools/docx2tex/",
+        "icon": "📝",
+    },
+]
+
+
+# EOF
