+++ apps/public_app/views/tools_data_text.py
@@ -1,0 +1,33 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""Research Tools data - Text domain tools."""
+
+from __future__ import annotations
+
+# Text tools
+TEXT_TOOLS = [
+    {
+        "name": "Markdown Renderer",
+        "description": "Real-time Markdown preview with syntax highlighting and table support.",
+        "use_case": "Format README files and documentation for data repositories",
+        "bookmarklet_url": "/tools/markdown-renderer/",
+        "icon": "📝",
+    },
+    {
+        "name": "Text Diff Checker",
+        "description": "Compare two text blocks side-by-side with highlighted differences.",
+        "use_case": "Compare dataset versions or track changes in results",
+        "bookmarklet_url": "/tools/text-diff-checker/",
+        "icon": "🔄",
+    },
+    {
+        "name": "JSON Formatter",
+        "description": "Format, validate, and beautify JSON data with syntax highlighting.",
+        "use_case": "Validate plot specifications and configuration files",
+        "bookmarklet_url": "/tools/json-formatter/",
+        "icon": "{ }",
+    },
+]
+
+
+# EOF
